+++ src/frontend/apps/impress/src/features/docs/doc-editor/components/comments/useComments.ts
@@ -3,6 +3,7 @@ import { useTranslation } from 'react-i18next';
 
 import { useCunninghamTheme } from '@/cunningham';
 import { User, avatarUrlFromName } from '@/features/auth';
+import { useEditorStore } from '@/features/docs/doc-editor/stores';
 import { Doc, useProviderStore } from '@/features/docs/doc-management';
 
 import { DocsThreadStore } from './DocsThreadStore';
@@ -16,6 +17,7 @@ export function useComments(
   const { provider } = useProviderStore();
   const { t } = useTranslation();
   const { themeTokens } = useCunninghamTheme();
+  const { setThreadStore } = useEditorStore();
 
   const threadStore = useMemo(() => {
     return new DocsThreadStore(
@@ -27,6 +29,18 @@ export function useComments(
       ),
     );
   }, [docId, canComment, provider?.awareness, user?.full_name]);
+
+  useEffect(() => {
+    if (canComment) {
+      setThreadStore(threadStore);
+    }
+
+    return () => {
+      if (canComment) {
+        setThreadStore(undefined);
+      }
+    };
+  }, [threadStore, setThreadStore, canComment]);
 
   useEffect(() => {
     return () => {